requirements_freeze.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
boto3==1.22.8
botocore==1.25.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.4
colorlog==6.6.0
cryptography==37.0.2
cycler==0.11.0
dask==2022.5.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
figrid==0.1.6
flake8==4.0.1
fonttools==4.33.3
fsspec==2022.3.0
globus-sdk==3.2.1
graphviz==0.20
ibl-neuropixel==0.1.0
ibllib==2.11.1
iblutil==1.0.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
locket==1.0.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mtscomp==1.0.2
nbclassic==0.3.7
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
notebook-shim==0.1.0
numba==0.55.1
numpy==1.21.0
ONE-api==1.11.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
patsy==0.5.2
phylib==2.4.1
pickleshare==0.7.5
Pillow==9.1.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
py==1.11.0
pyarrow==7.0.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.3.0
pynrrd==0.4.3
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
svgutils==0.3.4
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.0
wincertstore==0.2
zipp==3.8.0